Decoding 2D
Runs, but is relatively untested. Works with a basic dataset in a
notebook and runs on the full data. Next step will be to implement
more tests and plot some of these results to see over time whether
what we're doing is as expected, but it looks pretty promising
right now.