Co-occurrence with 2D tuning curves
Field detection and neuron assignment to different maze trajectories
is now done with 2D tuning curves.

Also, implemented a "multi-unit activity" threshold that only includes
the sharp-wave ripple events that also have a min number of neurons
spiking during that time.

This co-occurrence is now suited to combine sessions. This is done
using a weighted mean, so the number of sharp-wave ripple events that
participate is taken into account.

The SEM is not included, still need to figure out how to best do that
with a weighted average.