Refactoring co-occurrence to use with run.py
Doesn't seem to give the same output as before. The cached tuning
curves are different from what was originally used (here, using
all track times whereas before it was only during Phase 3). And R067d6
was added to the analysis, all other things should be the same, yet the
output is different. Investigate.