Correcting shortcut positions
Loads neuralynx targets and chooses the target with data if
both targets do not have data and if they do, then chooses the
target not contaminated by the feeder LED. Otherwise averages over
both targets.

This gets rid of the mini-mazes and leaves few "jumps" in the position
data - while keeping much of the position data originally recorded by
neuralynx.

Still need to write this into a function and incorporate it into the
shortcut workflow.