Reorganization to load data using vdmlab repository
Should be more reliable than using the *.mat intermediate, but still
includes a *.pkl intermediate for speedup.

Here the shortcut position was corrected using a different method.
Using the neuralynx target information, averaging when no corruption,
and taking non-feeder location when the targets disagree. This will
likely change the output of the analysis as more data points are included
in the analysis.