Raise speed threshold and restrict position for time bins
The old speed threshold (0.1) was allowing too many locations at the
feeder. Here, using 0.4 limits the positions at the feeder locations.

Here using restricted position time for generating time bins used
in decoding. This provides a more accurate representation of which
time bins accurately decode to a location since the spikes are also
limited to this time period. Otherwise, many positions did not contain
spikes and therefore resulted in lots of nans in the decoded array.