Plotting intermediate steps for decoding
And restricting tuning curves to only neurons that fire above a
mininum (here, 100) number of spikes during that phase. This could
be done in a way that makes it easier to filter neurons (eg. make
those tuning curves None or something) so that in later analysis steps
you're always using the same neurons (filtering the low firing ones out
of spikes).