Matching trials
Note: some sessions only have 3 novel trials because during the
experiment I ensured that the rats had a minimum of 5 passes over
each track segment. Since the novel trajectories end in a dead end
this meant that each "trial" counted as two passes through the
environment - one from start to the end and one from the end back.
Here I may need to account for that when equalizing the "trials"
between novel, u, and shortcut.